+++ contact/tool.js
@@ -1,0 +1,8 @@
+
+// Tool Javascript
+
+init.push(() => {
+    setInterval(() => {
+        
+    }, 250);
+});